+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - Bug #64: Update exception messages to use `SvgProperty` constants in attribute tests (@terabytesoftw)
 - Enh #65: Add `ui-awesome/html-mixin` package for reusable HTML attribute, mixins traits and update related imports accordingly (@terabytesoftw)
 - Bug #66: Better naming `SvgProperty` with `SvgAttribute` for SVG attribute constants and update related functionality and tests (@terabytesoftw)
+- Enh #67: Add `Image` class and corresponding tests for SVG `<image>` element functionality (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024